Conflicts:
	airtime_mvc/application/controllers/ApiController.php
	airtime_mvc/application/controllers/LibraryController.php
	airtime_mvc/application/controllers/PlaylistController.php
	airtime_mvc/application/models/Playlist.php
	python_apps/media-monitor/airtimefilemonitor/airtimemediamonitorbootstrap.py
	python_apps/media-monitor/airtimefilemonitor/airtimemetadata.py

+++ airtime_mvc/application/models/Playlist.php
@@ -453,13 +453,14 @@ class Application_Model_Playlist
             ->filterByDbPlaylistId($this->id)
             ->filterByDbPosition($pos)
             ->findOne();
+            
+        
 
-            #Propel returns values in form 00.000000 format which is for only seconds.
-            $fadeIn = $row->getDbFadein();
-            $fadeOut = $row->getDbFadeout();
-
-            return array($fadeIn, $fadeOut);
-    }
+        #Propel returns values in form 00.000000 format which is for only seconds.
+        $fadeIn = $row->getDbFadein();
+        $fadeOut = $row->getDbFadeout();
+        return array($fadeIn, $fadeOut);
+	}
 
     /**
      * Change fadeIn and fadeOut values for playlist Element
@@ -800,12 +801,33 @@ class Application_Model_Playlist
      * Delete playlists that match the ids..
      * @param array $p_ids
      */
-    public static function DeletePlaylists($p_ids)
+    public static function deletePlaylists($p_ids, $p_userId)
     {
-        CcPlaylistQuery::create()->findPKs($p_ids)->delete();
+        $leftOver = self::playlistsNotOwnedByUser($p_ids, $p_userId);
+        if (count($leftOver) == 0) {
+            CcPlaylistQuery::create()->findPKs($p_ids)->delete();
+        } else {
+            throw new PlaylistNoPermissionException;
+        }
+    }
+    
+    // This function returns that are not owen by $p_user_id among $p_ids 
+    private static function playlistsNotOwnedByUser($p_ids, $p_userId){
+        $ownedByUser = CcPlaylistQuery::create()->filterByDbCreatorId($p_userId)->find()->getData();
+        $selectedPls = $p_ids;
+        $ownedPls = array();
+        foreach ($ownedByUser as $pl) {
+            if (in_array($pl->getDbId(), $selectedPls)) {
+                $ownedPls[] = $pl->getDbId();
+            }
+        }
+        
+        $leftOvers = array_diff($selectedPls, $ownedPls);
+        return $leftOvers;
     }
 
 } // class Playlist
 
 class PlaylistNotFoundException extends Exception {}
+class PlaylistNoPermissionException extends Exception {}
 class PlaylistOutDatedException extends Exception {}
